Progress on java tests

+++ tests/practice_run_tests/java/solution_files/Sol12.java
@@ -1,11 +1,11 @@
 package tests.practice_run_tests.java.solution_files;
 
 public class Sol12 {
-  public static int solve(double num, double[] arr) {
+  public static long solve(double num, double[] arr) {
     double ret = num;
     for (double d : arr) {
       ret += d;
     }
-    return (int) ret;
+    return (long) ret;
   }
 }
