+++ tests/practice_run_tests/java/solution_files/Sol13.java
@@ -1,7 +1,7 @@
 package tests.practice_run_tests.java.solution_files;
 
 public class Sol13 {
-  public static double solve(int num) {
+  public static double solve(long num) {
     return ((double) num) / 2;
   }
 }
