+++ tests/practice_run_tests/java/solution_files/Sol6.java
@@ -1,7 +1,7 @@
 package tests.practice_run_tests.java.solution_files;
 
 public class Sol6 {
-  public static String solve(String a, int b) {
+  public static String solve(String a, long b) {
     return "foo";
   }
 }
